+++ packages/ckeditor5-core/tests/_utils/virtualtesteditor.js
@@ -33,14 +33,11 @@ export default class VirtualTestEditor extends Editor {
 
 			resolve(
 				editor.initPlugins()
+					.then( () => editor.data.init( config.initialData || '' ) )
 					.then( () => {
-						editor.data.init( config.initialData || '' );
-
-						// Fire `data#ready` event manually as `data#init()` method is not used.
-						editor.data.fire( 'ready' );
 						editor.fire( 'ready' );
+						return editor;
 					} )
-					.then( () => editor )
 			);
 		} );
 	}